+++ CHANGELOG.rst
@@ -6,6 +6,12 @@ TBR
 ---
 
 * Use the new ``web_poet.HttpResponse`` which replaces ``web_poet.ResponseData``.
+* Support for the new features in ``web_poet>=0.2.0`` for supporting additional 
+  requests inside Page Objects:
+
+    * Created new providers for ``web_poet.Meta`` and ``web_poet.HttpClient``.
+    * Using the said additional requests needs ``async/await`` support in
+      ``asyncio``. This raises the minimum scrapy requirement to ``scrapy>=2.6.0``.
 
 
 0.3.0 (2022-01-28)