+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ TBR
     * Created new providers for ``web_poet.Meta`` and ``web_poet.HttpClient``.
     * Using the said additional requests needs ``async/await`` support in
       ``asyncio``. This raises the minimum scrapy requirement to ``scrapy>=2.6.0``.
+* add ``async`` support for ``callback_for``.
 
 
 0.3.0 (2022-01-28)